==> notebook/2.Nueva creación de tablas.ipynb <==
import pandas as pd

import os

import sys
sys.path.append('../src')
from scrape_transform_save import *

# #PRIMERO TABLA DE CATEGORÍAS:

categorias = ['Esperanza de vida',
            'Morbilidad',
            'Mortalidad',
            'Estilo de Vida',]

categoría_dicc = {'Categoría': categorias,
            'Categoría_ID': [i for i in range(1,len(categorias)+1)]}

Categoría = pd.DataFrame(categoría_dicc)


Categoría.to_csv('../data/processed/Categoría.csv',index=False)

# SEGUNDO TABLA DE INDICADORES:

#OBTENEMOS NOMBRES DE INDICADORES DE LOS TÍTULOS DE LOS ARCHIVOS:

indicadores = []

for archivo in os.listdir('../data/raw'):
        
        if archivo.endswith('.xls'):
            
                file = os.path.join('../data/raw', archivo)
                
                indicadores.append((file.replace('.xls', '').replace('../data/raw\\','').replace('Anos','Años').replace('anos','años').replace('_',' ')))


indicadores_por_categoría = {'Esperanza de Vida': ['Años de vida saludable al nacer','Años de vida saludable a los 65 años','Esperanza de vida al nacer','Esperanza de vida a los 65 años','Indice de calidad de vida infantil relacionada con la salud'],
                            'Morbilidad': [ 'Incidencia de nuevos diagnósticos de VIH',
                                            'Incidencia de tuberculosis por 100,000 hab',
                                            'Prevalencia registrada de cardiopatía isquémica en población de 40 y más años',
                                            'Prevalencia registrada de diabetes mellitus en población general',
                                            'Prevalencia registrada de EPOC en población de 40 y más años',
                                            'Prevalencia registrada de hipertensión arterial en población general',
                                            'Prevalencia registrada de síndromes dolorosos de columna vertebral en población general',
                                            'Prevalencia registrada de trastornos mentales en población general',
                                            'Tasa ajustada de hospitalización por asma por 10,000 hab',
                                            'Tasa ajustada de hospitalización por diabetes mellitus por 10,000 hab',
                                            'Tasa ajustada de hospitalización por enfermedades cerebrovasculares por 10,000 hab',
                                            'Tasa ajustada de hospitalización por enfermedad hipertensiva por 10,000 hab',
                                            'Tasa ajustada de hospitalización por EPOC por 10,000 hab',
                                            'Tasa ajustada de hospitalización por infarto agudo de miocardio por 10,000 hab',
                                            'Tasa ajustada de hospitalización por Insuficiencia cardiaca congestiva por 10,000 hab',
                                            'Tasa bruta de hospitalización por enfermedades cerebrovasculares por 10,000 hab',
                                            'Tasa bruta de hospitalización por infarto agudo de miocardio por 10,000 hab',
                                            'Tasa de mortalidad ajustada por edad por cancer por 100,000 hab',
                                            'Porcentaje de recién nacidos a término con bajo peso',
                                            'Incidencia de accidentes de trabajo'],
                                'Mortalidad': ['Número de victimas mortales de accidentes de tráfico',
                                            'Numero de victimas no mortales de accidente de trafico',
                                            'Tasa bruta de hospitalización por enfermedades cerebrovasculares por 10,000 hab',
                                            'Tasa bruta de hospitalización por infarto agudo de miocardio por 10,000 hab',
                                            'Tasa de mortalidad ajustada por edad por enfermedad pulmonar obstructiva cronica por 100,000 hab',
                                            'Tasa de mortalidad ajustada por edad por neumonia e influenza por 100,000 hab',
                                            'Tasa de mortalidad ajustada por edad por suicidio por 100,000 hab',
                                            'Tasa de natalidad por 1000 habitantes',
                                            'Tasa de mortalidad ajustada por edad por cancer por 100,000 hab',
                                            'Tasa de mortalidad ajustada por edad por cardiopatia isquemica por 100,000 hab',
                                            'Tasa de mortalidad ajustada por edad por diabetes mellitus por 100,000 hab',
                                            'Tasa de mortalidad ajustada por edad por enfermedad cerebrovascular por 100,000 hab',
                                            'Tasa de mortalidad ajustada por edad por enfermedad cronica del higado por 100,000 hab',
                                            'Tasa de mortalidad ajustada por edad por enfermedad pulmonar obstructiva cronica por 100,000 hab',
                                            'Tasa de mortalidad ajustada por edad por neumonia e influenza por 100,000 hab',
                                            'Tasa de mortalidad ajustada por edad por suicidio por 100,000 hab',
                                            'Tasa de mortalidad prematura por cáncer ajustada por edad por 100,000 hab',
                                            'Tasa de mortalidad prematura por diabetes mellitus ajustada por edad por 100,000 hab',
                                            'Tasa de mortalidad prematura por enfermedad vascular cerebral ajustada por edad por 100,000 hab',
                                            'Tasa de mortalidad prematura por EPOC ajustada por edad por 100,000 hab',
                                            'Tasa de mortalidad prematura por cardiopatia isquemica ajustada por edad por 100 000 hab',
                                            'Tasa de mortalidad infantil por 1000 nacidos vivos'],
                                'Estilo de Vida': [ 'Consumo diario de frutas yo verduras en adultos',
                                            'Prevalencia de consumo de tabaco en población adulta',
                                            'Prevalencia de consumo diario de tabaco en jóvenes',
                                            'Prevalencia de obesidad en población de 18 y más años',
                                            'Prevalencia de obesidad en población infantil',
                                            'Prevalencia de sedentarismo en jóvenes',
                                            'Prevalencia de sedentarismo en población adulta',
                                            'Consumo diario de frutas yo verduras en jovenes',
                                            'Consumo intensivo de alcohol al menos una vez al mes en los ultimos 12 meses en jóvenes',
                                            'Porcentaje de bebedores de riesgo de alcohol',
                                            'Porcentaje de consumidores de cannabis en escolares adolescentes de 14 a 18 años',
                                            'Porcentaje de la población que recibe suministro con Cobertura del Sistema Nacional de Aguas de Consumo',
                                            'Porcentaje de la población que declara escasez de zonas verdes en el entorno de su vivienda',
                                            'Porcentaje de la población que declara que el ruido procedente del exterior de su vivienda es molesto',
                                            'Porcentaje de población adulta en riesgo de mala salud mental',
                                            'Porcentaje de población con valoración positiva de su estado de salud',
                                            'Porcentaje de población de 65 años y más años con valoración positiva de su estado de salud',
                                            'Porcentaje de población de clase social baja',
                                            'Porcentaje de poblacion que recibe suministro con Cobertura del Sistema Nacional de Aguas de Consumo',
                                            'Proporcion de poblacion de 25 a 64 años con nivel de estudios de 1 Etapa de educacion secundaria o inferior',
                                            'Tasa de desempleo en menores de 25 años',
                                            'Tasa de desempleo total',
                                            'Dificultad para la movilidad en 65 y mas años',
                                            'Deterioro cognitivo en 65 y mas años',
                                            'Indice de dependencia en 65 y mas años',
                                            'Indice de dependencia',
                                            'Indice de dependencia juvenil',
                                            'Ingresos medios por persona euros',
                                            'Tasa de riesgo de pobreza indicador AROPE',
                                            'Tasa de riesgo de pobreza indicador AROPE',
                                            'Limitacion para la realizacion de actividades relacionadas con el hogar en 65 y mas años']}

indicadores_por_categoría2 = {}

for key, values in indicadores_por_categoría.items():
    new_values = []
    for value in values:
        new_value = value.replace('Anos', 'años').replace('anos', 'años').replace('á', 'a').replace('é', 'e').replace('í', 'i').replace('ó', 'o').replace('ú', 'u').replace(',',' ')
        new_values.append(new_value)
    indicadores_por_categoría2[key] = new_values
    
Indicadores = pd.DataFrame({'Indicadores_ID':[i for i in range(1,len(indicadores)+1)],'Indicadores':indicadores})

diccionario_inverso = {indicador: categoria for categoria,
                       indicadores in indicadores_por_categoría2.items() for indicador in indicadores}

Indicadores['Categoría_ID'] = Indicadores['Indicadores'].map(diccionario_inverso)

diccionario_inverso2 = {'Esperanza de Vida': '1',
 'Morbilidad': '2',
 'Mortalidad': '3',
 'Estilo de Vida': '4'}

Indicadores['Categoría_ID'] = Indicadores['Categoría_ID'].map(diccionario_inverso2)

Indicadores.to_csv('../data/processed/Indicadores.csv',index=False)


Indicadores

Indicadores.Indicadores.unique()

#Creamos la tabla de Region_ID

comunidades_dict = {'España (ES)': 1, 
    'Andalucía (AN)': 2,
    'Aragón (AR)': 3,
    'Asturias, Principado de (AS)': 4,
    'Balears, Illes (IB)': 5,
    'Canarias (CN)': 6,
    'Cantabria (CB)': 7,
    'Castilla y León (CL)': 8,
    'Castilla - La Mancha (CM)': 9,
    'Cataluña (CT)': 10,
    'Comunitat Valenciana (VC)': 11,
    'Extremadura (EX)': 12,
    'Galicia (GA)': 13,
    'Madrid, Comunidad de (MD)': 14,
    'Murcia, Región de (MC)': 15,
    'Navarra, C. Foral de (NC)': 16,
    'País Vasco (PV)': 17,
    'La Rioja (RI)': 18,
    'Ceuta (CE)': 19,
    'Melilla (ML)': 20}

Region = pd.DataFrame(list(comunidades_dict.items()), columns=['Region', 'Region_ID'])
Region.to_csv('../data/processed/Region.csv', index=False)

# CREAMOS TABLA DE DATOS/VALORES:

tabla = pd.read_excel('../data/raw/Anos_de_vida_saludable_a_los_65_anos.xls')

tabla = clean_columns(tabla)

tabla = create_df(tabla)

# VALE AHORA TENGO QUE HACER LO QUE SE HACE CON YEAR CON GÉNERO:

tabla.head()

df_largo = tabla.melt(id_vars=['Region', 'Año'], value_vars=['Hombre', 'Mujer','Total'], var_name='Genero', value_name='Valores')

# melt, coge Region y Año y los mantiene y Hombre y Mujer los 'derrite' para unificarlos con las columnas que mantenemos 💞

# Ahora, tengo las columnas Region, Año,Género y valores Ok. Queda añadir Dato_ID,Indicador_ID y Region_ID
df_largo.head()


from sqlalchemy import column


def process_excel_files_test(Source_folder, Destination_folder):
    '''
    Comentario
    '''
    # Lista para almacenar los DataFrames procesados y luego concatenar
    
    dfs = []
    
    for filename in os.listdir(Source_folder):
        
        if filename.endswith(".xls"): 
            
            try:    
                
                df = pd.read_excel(os.path.join(Source_folder, filename))
                
                df = clean_columns(df)
                
                df = create_df(df)
                            
                    
                if all('Hombre' not in col for col in df.columns):
                    
                    try:
                        
                        df = df[df['Region'].notna()]
                            
                        df = df[df.Total.notna()]
                        
                        df = df.melt(id_vars=['Region', 'Año'], value_vars=['Total'], var_name='Genero', value_name='Valores')
                        
                        df['Indicadores_ID'] = os.path.splitext(f'{filename}')[0]
                        dfs.append(df)
                        
                    except Exception as e :
                        
                        print(f'El archivo{file},dio error:{e}')
                    
                else:   
                    df = df.drop(columns = 'Total')
                    df = df[df['Region'].notna()]
                                                                    
                    df = df.melt(id_vars=['Region', 'Año'], value_vars=['Hombre', 'Mujer'], var_name='Genero', value_name='Valores')
                                        
                    df['Indicadores_ID'] = os.path.splitext(f'{filename}')[0]

                    df = df.apply(pd.to_numeric, downcast='integer', errors='ignore')
                        
                    df = df[df.Valores.notna()]
                    
                    dfs.append(df)
                        
            except Exception as e :
                    
                    print(f'El archivo{file},dio error:{e}')
                
            continue
     
    final_df = pd.concat(dfs, ignore_index=True)
    
    comunidades_dict = {'España (ES)': 1, 
    'Andalucía (AN)': 2,
    'Aragón (AR)': 3,
    'Asturias, Principado de (AS)': 4,
    'Balears, Illes (IB)': 5,
    'Canarias (CN)': 6,
    'Cantabria (CB)': 7,
    'Castilla y León (CL)': 8,
    'Castilla - La Mancha (CM)': 9,
    'Cataluña (CT)': 10,
    'Comunitat Valenciana (VC)': 11,
    'Extremadura (EX)': 12,
    'Galicia (GA)': 13,
    'Madrid, Comunidad de (MD)': 14,
    'Murcia, Región de (MC)': 15,
    'Navarra, C. Foral de (NC)': 16,
    'País Vasco (PV)': 17,
    'La Rioja (RI)': 18,
    'Ceuta (CE)': 19,
    'Melilla (ML)': 20}

    final_df['Region'] =  final_df['Region'].map(comunidades_dict)    

    final_df['Valores_ID'] = [e for e in range(1,len(final_df)+1)]

    final_df.to_csv(os.path.join(Destination_folder, 'Valores.csv'), index=False)
    
    return final_df


df = process_excel_files_test('../data/raw', '../data/processed')

df = pd.read_csv('../data/processed/Valores.csv')

df['Indicadores_ID'] = df['Indicadores_ID'].apply(lambda x: x.replace('Anos', 'Años').replace('anos', 'años').replace('_',' '))

df

# Supongamos que 'df' es tu DataFrame y 'tabla_indicadores' es la tabla Indicadores
# Primero, creamos un diccionario que mapea los valores actuales de 'Indicadores_ID' a los nuevos valores
mapeo = dict(Indicadores[['Indicadores', 'Indicadores_ID']].values)

# Luego, usamos la función map() para reemplazar los valores en 'Indicadores_ID'
df['Indicadores_ID'] = df['Indicadores_ID'].map(mapeo)


df.columns = ['Unnamed: 0', 'Region_ID', 'Año', 'Genero', 'Valores', 'Indicadores_ID',
       'Valores_ID']

df.drop(columns=df.columns[0],inplace=True)

df

df.to_csv('../data/processed/Valores.csv',index=False)
