==> notebook/3.Load and create DB.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)

import os

import mysql.connector
from sqlalchemy import create_engine, text

import warnings
warnings.filterwarnings('ignore')

import sys
sys.path.append('../src')
from sql_upload_manager import *

usser = 'root'
key = open('password.txt','r').read()
host = 'localhost'

conexion, cursor = conection_to_SQL('root', key, 'localhost')

database = 'IndiSalud'

cursor.execute(f'CREATE DATABASE IF NOT EXISTS {database}')


for filename in os.listdir('../data/processed'):
        
    if filename != "Valores.csv" and filename !='Indicadores.csv':
        
        df = pd.read_csv(f'../data/processed/{filename}')
               
        engine = create_engine(f'mysql+mysqlconnector://root:{key}@localhost:3306/{database}', echo=True)
        filename = filename.split('.')[0]
        df.to_sql(f'{filename}',
                  con=engine,
                  if_exists='append',
                  index=False)
        
        for column in df.columns:
            if column.endswith('_ID'):
                
                with engine.connect() as con:
                    
                    con.execute(text(f'ALTER TABLE `{filename}` ADD PRIMARY KEY (`{column}`)'))
                    print(f"The table {filename} has been successfully loaded into SQL.")
                    
                    
# INDICADORES:

df = pd.read_csv('../data/processed/Indicadores.csv')
engine = create_engine(f'mysql+mysqlconnector://root:{key}@localhost:3306/{database}', echo=True)
df.to_sql('Indicadores', con=engine, if_exists='append', index=False)
        
with engine.connect() as con:
    con.execute(text('ALTER TABLE `Indicadores` ADD PRIMARY KEY (`Indicadores_ID`),ADD FOREIGN KEY (`Categoría_ID`) REFERENCES `Categoría` (`Categoría_ID`)'))
        
    
# VALORES :

df = pd.read_csv('../data/processed/Valores.csv')
engine = create_engine(f'mysql+mysqlconnector://root:{key}@localhost:3306/{database}', echo=True)
df.to_sql('Valores', con=engine, if_exists='append', index=False)
        
with engine.connect() as con:
    con.execute(text('ALTER TABLE `Valores` ADD CONSTRAINT fk1_regionid FOREIGN KEY (`Region_ID`) REFERENCES `Region` (`Region_ID`), ADD CONSTRAINT fk2_indicadores FOREIGN KEY (`Indicadores_ID`) REFERENCES `Indicadores` (`Indicadores_ID`)'))
